set up property based test to check multiple queries per data generation
