+++ saffron/Cargo.toml
@@ -38,9 +38,8 @@ time = { version = "0.3", features = ["macros"] }
 tracing = "0.1"
 tracing-subscriber = { version = "0.3", features = [ "ansi", "env-filter", "fmt", "time" ] }
 
-ark-std.workspace = true
-
 [dev-dependencies]
+ark-std.workspace = true
 ctor = "0.2"
 proptest.workspace = true
 once_cell.workspace = true
